Convert files to dirty typescript and disable action modal outside click
