+++ frontend/src/metabase/components/Modal/WindowModal.tsx
@@ -1,0 +1,133 @@
+import React, { Component, CSSProperties, ReactNode } from "react";
+import { CSSTransition, TransitionGroup } from "react-transition-group";
+import cx from "classnames";
+import {
+  getModalContent,
+  ModalSize,
+  modalSizes,
+  ModalWrapper,
+} from "metabase/components/Modal/common";
+
+import SandboxedPortal from "metabase/components/SandboxedPortal";
+
+export type WindowModalProps = {
+  isOpen?: boolean;
+  onClose?: () => void;
+  children?: ReactNode;
+  className?: string;
+  backdropClassName?: string;
+  fullPageModal?: boolean;
+  enableMouseEvents?: boolean;
+  enableTransition?: boolean;
+  noOnClickOutsideWrapper?: boolean;
+  noBackdrop?: boolean;
+  noCloseOnBackdrop?: boolean;
+  form?: unknown;
+  formModal?: boolean;
+  style?: CSSProperties;
+} & {
+  [size in ModalSize]?: boolean;
+};
+
+export class WindowModal extends Component<WindowModalProps> {
+  _modalElement: HTMLDivElement;
+
+  static defaultProps = {
+    className: "Modal",
+    backdropClassName: "Modal-backdrop",
+    enableTransition: true,
+  };
+
+  constructor(props: WindowModalProps) {
+    super(props);
+
+    this._modalElement = document.createElement("div");
+    this._modalElement.className = "ModalContainer";
+    document.body.appendChild(this._modalElement);
+  }
+
+  componentWillUnmount() {
+    if (this._modalElement.parentNode) {
+      this._modalElement.parentNode.removeChild(this._modalElement);
+    }
+  }
+
+  handleDismissal = () => {
+    if (this.props.onClose) {
+      this.props.onClose();
+    }
+  };
+
+  _modalComponent() {
+    const className = cx(
+      this.props.className,
+      ...modalSizes
+        .filter(type => this.props[type])
+        .map(type => `Modal--${type}`),
+    );
+    return (
+      <ModalWrapper
+        backdropElement={this._modalElement}
+        handleDismissal={this.handleDismissal}
+        {...this.props}
+      >
+        <div
+          className={cx(className, "relative bg-white rounded")}
+          role="dialog"
+        >
+          {getModalContent({
+            ...this.props,
+            fullPageModal: false,
+            // if there is a form then its a form modal, or if there's a form
+            // modal prop use that
+            formModal: !!this.props.form || this.props.formModal,
+          })}
+        </div>
+      </ModalWrapper>
+    );
+  }
+
+  render() {
+    const {
+      enableMouseEvents,
+      backdropClassName,
+      isOpen,
+      style,
+      enableTransition,
+    } = this.props;
+    const backdropClassnames =
+      "flex justify-center align-center fixed top left bottom right";
+
+    return (
+      <SandboxedPortal
+        container={this._modalElement}
+        enableMouseEvents={enableMouseEvents}
+      >
+        <TransitionGroup
+          appear={enableTransition}
+          enter={enableTransition}
+          exit={enableTransition}
+        >
+          {isOpen && (
+            <CSSTransition
+              key="modal"
+              classNames="Modal"
+              timeout={{
+                appear: 250,
+                enter: 250,
+                exit: 250,
+              }}
+            >
+              <div
+                className={cx(backdropClassName, backdropClassnames)}
+                style={style}
+              >
+                {this._modalComponent()}
+              </div>
+            </CSSTransition>
+          )}
+        </TransitionGroup>
+      </SandboxedPortal>
+    );
+  }
+}
